use orjson instead
