+++ testplan/common/utils/json.py
@@ -1,32 +1,31 @@
 import json
+from pathlib import Path
+from typing import Union
 
-_USE_RAPIDJSON = False
+_USE_ORJSON = False
 
 try:
-    import rapidjson
+    import orjson
 except ImportError:
     pass
 else:
-    _USE_RAPIDJSON = True
+    _USE_ORJSON = True
 
 
 def json_loads(data: str):
-    if _USE_RAPIDJSON:
-        # being explicit here
-        return rapidjson.loads(data, number_mode=rapidjson.NM_NAN)
+    if _USE_ORJSON:
+        return orjson.loads(data)
     else:
         return json.loads(data)
 
 
 def json_dumps(data, indent_2=False, default=None) -> str:
-    if _USE_RAPIDJSON:
-        # being explicit here
-        return rapidjson.dumps(
+    if _USE_ORJSON:
+        return orjson.dumps(
             data,
-            indent=2 if indent_2 else None,
             default=default,
-            number_mode=rapidjson.NM_NAN,
-        )
+            option=orjson.OPT_INDENT_2 if indent_2 else 0,
+        ).decode()
     else:
         if default:
 
@@ -37,3 +36,8 @@ def json_dumps(data, indent_2=False, default=None) -> str:
         else:
             _E = None
         return json.dumps(data, cls=_E, indent=2 if indent_2 else None)
+
+
+def json_load_from_path(path: Union[str, Path]) -> dict:
+    with open(path) as fp:
+        return json_loads(fp.read())
