+++ testplan/report/testing/schemas.py
@@ -1,23 +1,22 @@
 """Schema classes for test Reports."""
 
 import functools
-import json
 
 from boltons.iterutils import is_scalar, remap
 from marshmallow import Schema, fields, post_load
 from marshmallow.utils import EXCLUDE
 
+from testplan.common.report import RuntimeStatus, Status
 from testplan.common.report.schemas import (
-    ReportSchema,
     BaseReportGroupSchema,
-    ReportLogSchema,
-    TimerField,
     ReportLinkSchema,
+    ReportLogSchema,
+    ReportSchema,
+    TimerField,
 )
-
-from testplan.common.report import Status, RuntimeStatus
 from testplan.common.serialization import fields as custom_fields
 from testplan.common.serialization.schemas import load_tree_data
+from testplan.common.utils.json import json_dumps, json_loads
 from testplan.report.testing.base import (
     TestCaseReport,
     TestGroupReport,
@@ -49,30 +48,33 @@ class EntriesField(fields.Field):
     Handle encoding problems gracefully
     """
 
-    # @staticmethod
-    # def _json_serializable(v):
-    #     try:
-    #         json.dumps(v, ensure_ascii=True)
-    #     except (UnicodeDecodeError, TypeError):
-    #         return False
-    #     else:
-    #         return True
+    @staticmethod
+    def _json_serializable(v):
+        try:
+            # for nan, inf & -inf,
+            # so they will be converted to str
+            if v != json_loads(json_dumps(v)):
+                return False
+        except (UnicodeDecodeError, TypeError):
+            return False
+        else:
+            return True
 
-    # def _serialize(self, value, attr, obj, **kwargs):
-    #     # we don't need a _deserialize() here as we don't (and can't)
-    #     # convert str back to non-json-serializable.
-    #     def visit(parent, key, _value):
-    #         """
-    #         return
-    #             True - keep the node unchange
-    #             False - remove the node
-    #             tuple - update the node data.
-    #         """
-    #         if is_scalar(_value) and not self._json_serializable(_value):
-    #             return key, str(_value)
-    #         return True
+    def _serialize(self, value, attr, obj, **kwargs):
+        # we don't need a _deserialize() here as we don't (and can't)
+        # convert str back to non-json-serializable.
+        def visit(parent, key, _value):
+            """
+            return
+                True - keep the node unchange
+                False - remove the node
+                tuple - update the node data.
+            """
+            if is_scalar(_value) and not self._json_serializable(_value):
+                return key, str(_value)
+            return True
 
-    #     return remap(value, visit=visit)
+        return remap(value, visit=visit)
 
 
 class TestCaseReportSchema(ReportSchema):
